+++ packages/amplify_api/ios/Classes/OperationsManager.swift
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License").
  * You may not use this file except in compliance with the License.
@@ -16,29 +16,29 @@
 import Foundation
 import Amplify
 
-public class OperationsManager{
+public class OperationsManager {
     /// Prevents concurrent access to arrays.
     private static let queue = DispatchQueue(label: "com.amazonaws.flutter.OperationsManager.concurrency")
-    
+
     /// Maps `cancelToken` to [Operation].
     private static var operationsMap: [String: Operation] = [:]
-    
+
     /// Maps [URLSessionTaskBehavior.taskBehaviorIdentifier] to [RESTOperation.id].
     private static var operationIdsByTaskId: [Int: UUID] = [:]
-    
+
     /// Maps [RESTOperation.id] to [URLResponse].
     private static var operationsResponseMap: [UUID: HTTPURLResponse] = [:]
-    
+
     public static func containsOperation(cancelToken: String) -> Bool {
         return operationsMap[cancelToken] != nil
     }
-    
+
     public static func addOperation(cancelToken: String, operation: Operation) {
         queue.sync {
             operationsMap[cancelToken] = operation
         }
     }
-    
+
     public static func removeOperation(cancelToken: String) {
         guard let operation = operationsMap[cancelToken] else {
             return
@@ -50,7 +50,7 @@ public class OperationsManager{
             }
         }
     }
-    
+
     public static func setTaskId(for cancelToken: String, taskId: Int) {
         guard let operation = operationsMap[cancelToken] as? RESTOperation else {
             return
@@ -59,7 +59,7 @@ public class OperationsManager{
             operationIdsByTaskId[taskId] = operation.id
         }
     }
-    
+
     public static func updateProgress(for taskId: Int, urlResponse: HTTPURLResponse) {
         guard let operationId = operationIdsByTaskId[taskId] else {
             return
@@ -68,15 +68,15 @@ public class OperationsManager{
             operationsResponseMap[operationId] = urlResponse
         }
     }
-    
+
     public static func getResponse(for cancelToken: String) -> HTTPURLResponse? {
         guard let operation = operationsMap[cancelToken] as? RESTOperation else {
             return nil
         }
         return operationsResponseMap[operation.id]
     }
-    
-    public static func cancelOperation(cancelToken : String){
+
+    public static func cancelOperation(cancelToken: String) {
         operationsMap[cancelToken]?.cancel()
         removeOperation(cancelToken: cancelToken)
     }