+++ packages/amplify_auth_cognito/CHANGELOG.md
@@ -1,3 +1,29 @@
+## 0.2.4 (2021-09-10)
+
+### Fixes
+
+- fix: CocoaPods relative import
+
+## 0.2.3 (2021-09-09)
+
+### Features
+
+- feat(auth): OIDC/Lambda Support (#777)
+- feat(auth): add options to updateAttribute, updateAttributes, resendUserAttributeConfirmationCode (#775)
+- feat(auth): add support for options to resetPassword, confirmResetPassword (#743)
+- feat(auth): add options to resendSignUpCode (#738)
+
+### Chores
+
+- chore: upgrade amplify-android 1.24.1 (#829)
+
+## 0.2.2 (2021-08-04)
+
+### Features
+
+- feat: enables clientMetadata for signUp (#713)
+- feat: Auth Devices API (#735)
+
 ## 0.2.1 (2021-07-27)
 
 ### Bug Fixes
@@ -134,6 +160,7 @@
 ## 0.0.2-dev.1 (2021-01-25)
 
 ### Features
+
 - feature: Amplify Auth Hosted UI (#309)
 
 ### Chores